Disable ESLint error in test TS type code
There's been an unexpected change in a dependency, so that . is now
consdered to be a bare name that's missing an extension. That's not
correct - we want to import the whole package here in this case. Since
this is only typing test code, it seems clearest to just disable the
rule for this one line here for now.